Parted the UI into components & pages
Changes;

Created components and then fixed it to display the "pages" instead
