+++ frontend/src/components/UserMovieList.jsx
