+++ frontend/src/components/WelcomePage.jsx
