+++ frontend/src/pages/AlgorithmPage.jsx
@@ -1,0 +1,9 @@
+/* eslint-disable react/prop-types */
+
+import Algorithm from '../components/Algorithm.jsx';
+
+const AlgorithmPage = ({ genres }) => {
+	return <Algorithm genres={genres} />;
+};
+
+export default AlgorithmPage;
