+++ frontend/src/pages/HomePage.jsx
@@ -1,0 +1,9 @@
+/* eslint-disable react/prop-types */
+
+import Home from '../components/Home.jsx';
+
+const HomePage = ({ genres }) => {
+	return <Home genres={genres} />;
+};
+
+export default HomePage;
