+++ frontend/src/pages/WelcomePage.jsx
@@ -1,0 +1,9 @@
+/* eslint-disable react/prop-types */
+
+import WelcomePage from '../components/WelcomePage.jsx';
+
+const Welcome = ({ setDisplay }) => {
+	return <WelcomePage setDisplay={setDisplay} />;
+};
+
+export default Welcome;
